requirements/base.txt
Django==1.7.6
psycopg2>=2.5.2
django-compressor>=1.4
django-modelcluster>=0.4
wagtail>=0.8.4
django-taggit>=0.12.2
django-libsass>=0.2
django-storages>=1.1.8
boto>=2.36.0
